requirements.txt
requests==2.31.0
tqdm>=4.66.0
PyYAML>=6.0.1
feedparser==6.0.11
arxiv>=2.1.0
beautifulsoup4==4.12.3
# openai<1.55 与 httpx>=0.28 不兼容（Client 不再接受 proxies=）
httpx>=0.27,<0.28
openai==1.30.0
markdown==3.6
python-dateutil==2.9.0
python-dotenv>=1.0.0